+++ apps/src/templates/teacherDashboard/OwnedSections.jsx
@@ -19,6 +19,8 @@ import styleConstants from '@cdo/apps/styleConstants';
 import AddSectionDialog from './AddSectionDialog';
 import EditSectionDialog from './EditSectionDialog';
 import SetUpSections from '../studioHomepages/SetUpSections';
+import {recordOpenEditSectionDetails} from './sectionHelpers';
+import experiments from '@cdo/apps/util/experiments';
 
 const styles = {
   button: {
@@ -56,11 +58,31 @@ class OwnedSections extends React.Component {
     viewHidden: false
   };
 
+  constructor(props) {
+    super(props);
+    this.onEditSection = this.onEditSection.bind(this);
+  }
+
   componentDidMount() {
     const {queryStringOpen, beginImportRosterFlow} = this.props;
 
     if (queryStringOpen === 'rosterDialog') {
       beginImportRosterFlow();
+    }
+  }
+
+  onEditSection(id) {
+    this.props.beginEditingSection(id);
+    if (experiments.isEnabled(experiments.TEACHER_DASHBOARD_SECTION_BUTTONS)) {
+      recordOpenEditSectionDetails(
+        id,
+        'owned_sections_table_with_dashboard_header_buttons'
+      );
+    } else {
+      recordOpenEditSectionDetails(
+        id,
+        'owned_sections_table_without_dashboard_header_buttons'
+      );
     }
   }
 
@@ -78,7 +100,6 @@ class OwnedSections extends React.Component {
       sectionIds,
       hiddenSectionIds,
       asyncLoadComplete,
-      beginEditingSection,
       locale
     } = this.props;
     const {viewHidden} = this.state;
@@ -98,7 +119,7 @@ class OwnedSections extends React.Component {
             {visibleSectionIds.length > 0 && (
               <OwnedSectionsTable
                 sectionIds={visibleSectionIds}
-                onEdit={beginEditingSection}
+                onEdit={this.onEditSection}
               />
             )}
             <div style={styles.buttonContainer}>
@@ -123,7 +144,7 @@ class OwnedSections extends React.Component {
                 </div>
                 <OwnedSectionsTable
                   sectionIds={hiddenSectionIds}
-                  onEdit={beginEditingSection}
+                  onEdit={this.onEditSection}
                 />
               </div>
             )}
