+++ apps/src/templates/teacherDashboard/TeacherDashboard.jsx
@@ -4,7 +4,9 @@ import {Route, Switch} from 'react-router-dom';
 import TeacherDashboardNavigation, {
   TeacherDashboardPath
 } from './TeacherDashboardNavigation';
+import experiments from '@cdo/apps/util/experiments';
 import TeacherDashboardHeader from './TeacherDashboardHeader';
+import TeacherDashboardHeaderWithButtons from './TeacherDashboardHeaderWithButtons';
 import StatsTableWithData from './StatsTableWithData';
 import SectionProgress from '@cdo/apps/templates/sectionProgress/SectionProgress';
 import ManageStudents from '@cdo/apps/templates/manageStudents/ManageStudents';
@@ -16,6 +18,21 @@ import EmptySection from './EmptySection';
 import _ from 'lodash';
 import firehoseClient from '../../lib/util/firehose';
 
+function Header(props) {
+  if(experiments.isEnabled(experiments.TEACHER_DASHBOARD_SECTION_BUTTONS)) {
+    return (
+      <div>
+        {/* TeacherDashboardNavigation must be outside of
+        TeacherDashboardHeader. Routing components do not work with
+        components using Connect/Redux. Library we could use to fix issue:
+        https://github.com/supasate/connected-react-router */}
+        <TeacherDashboardHeaderWithButtons/> <TeacherDashboardNavigation/>
+      </div>
+    )
+  } else {
+    return <TeacherDashboardHeader sectionName={props.sectionName} />
+  }
+}
 class TeacherDashboard extends Component {
   static propTypes = {
     studioUrlPrefix: PropTypes.string.isRequired,
@@ -49,6 +66,7 @@ class TeacherDashboard extends Component {
     }
   }
 
+
   render() {
     const {
       location,
@@ -76,15 +94,7 @@ class TeacherDashboard extends Component {
 
     return (
       <div>
-        {includeHeader && (
-          <div>
-            {/* TeacherDashboardNavigation must be outside of
-            TeacherDashboardHeader. Routing components do not work with
-            components using Connect/Redux. Library we could use to fix issue:
-            https://github.com/supasate/connected-react-router */}
-            <TeacherDashboardHeader /> <TeacherDashboardNavigation />
-          </div>
-        )}
+        {includeHeader && (<Header {... this.props}/>)}
         <Switch>
           <Route
             path={TeacherDashboardPath.manageStudents}
