+++ apps/src/templates/teacherDashboard/urlHelpers.js
@@ -1,8 +1,4 @@
 const dashboardPrefix = '/teacher_dashboard/sections/';
-import {navigateToHref} from '@cdo/apps/utils';
-import {TeacherDashboardPath} from '@cdo/apps/templates/teacherDashboard/TeacherDashboardNavigation';
-import firehoseClient from '../../lib/util/firehose';
-import _ from 'lodash';
 
 /**
  * Returns the URL in teacher dashboard given a section id and (optional) path.
@@ -24,26 +20,3 @@ export const scriptUrlForStudent = (sectionId, scriptName, studentId) => {
   return `/s/${scriptName}?section_id=${sectionId}&user_id=${studentId}&viewAs=Teacher`;
 };
 
-export const switchToSection = (toSectionId, fromSectionId) => {
-  const baseUrl = `/teacher_dashboard/sections/${toSectionId}/`;
-  const currentTab = _.last(_.split(window.location.pathname, '/'));
-  const teacherNavigationTabs = _.values(TeacherDashboardPath);
-  const sectionUrl = _.includes(teacherNavigationTabs, `/${currentTab}`)
-    ? baseUrl.concat(currentTab)
-    : baseUrl;
-  navigateToHref(sectionUrl);
-
-  firehoseClient.putRecord(
-    {
-      study: 'teacher_dashboard_actions',
-      study_group: currentTab,
-      event: 'change_section',
-      data_json: JSON.stringify({
-        section_id: fromSectionId,
-        old_section_id: fromSectionId,
-        new_section_id: toSectionId
-      })
-    },
-    {includeUserId: true}
-  );
-};
